requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pymysql
psycopg2-binary
pymssql
